backend/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
passlib[bcrypt]
PyJWT
